Java Masterclass/src/S18_Debugging_JUnit/Banking/BankAccount.java
package S18_Debugging_JUnit.Banking;


public class BankAccount {
    private String firstName;
    private String lastName;
    private double balance;

    public static final int CHECKING = 1;
    public static final int SAVINGS = 2;

    private int accountType;


    public BankAccount(String firstName, String lastName, double balance, int accountType) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.balance = balance;
        this.accountType = accountType;
    }

    public double deposit(double deposit, boolean branch) {
        this.balance += deposit;
        return balance;
    }

    public double withdraw(double withdraw, boolean branch) {
        balance -= withdraw;
        return balance;
    }

    public double getBalance() {
        return balance;
    }

    public boolean isChecking() {
        return accountType == CHECKING;
    }


}
